feat: update model of user
