+++ server/src/resources/user/user.router.js
@@ -1,28 +1,72 @@
-const router = require('express').Router()
-const userService = require('./user.service')
+const express = require('express')
+const router = express.Router()
+const bcrypt = require('bcryptjs')
+const jwt = require('jsonwebtoken')
+const { check, validationResult } = require('express-validator')
+
 const User = require('./user.model')
 
-router.route('/').get(async (req, res) => {
-    const users = await userService.getAll()
-    if (!users) {
-        res.status('404')
-    }
-    res.json(users.map(User.toResponse))
-})
+router.post(
+    '/',
+    [
+        check('name', 'Name is required').not().isEmpty(),
+        check('username', 'Username is required').not().isEmpty(),
+        check('email', 'Please include a valid email').isEmail(),
+        check(
+            'password',
+            'Please enter a password with 6 or more characters'
+        ).isLength({ min: 6 }),
+    ],
+    async (req, res) => {
+        const errors = validationResult(req)
+        if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() })
+        }
 
-router.route('/:id').get(async (req, res) => {
-    const user = await (
-        await userService.getUserById(req.params.id)
-    ).populate('album')
-    if (!user) {
-        res.status('404')
-    }
-    res.json(User.toResponse(user))
-})
+        const { name, username, email, password } = req.body
 
-router.route('/').post(async (req, res) => {
-    const user = await userService.createUser(req.body)
-    res.status(200).json(User.toResponse(user))
-})
+        try {
+            let user = await User.findOne({ email })
+
+            if (user) {
+                return res
+                    .status(400)
+                    .json({ errors: [{ msg: 'User already exists' }] })
+            }
+
+            user = new User({
+                name,
+                username,
+                email,
+                password,
+            })
+
+            const salt = await bcrypt.genSalt(10)
+
+            user.password = await bcrypt.hash(password, salt)
+
+            await user.save()
+
+            const payload = {
+                user: {
+                    id: user.id,
+                },
+            }
+
+            jwt.sign(
+                payload,
+                process.env.JWT_SECRET_KEY,
+                { expiresIn: 360000 },
+                (err, token) => {
+                    if (err) throw err
+                    res.json({ token })
+                }
+            )
+        } catch (err) {
+            console.error(err.message)
+            res.status(500).send('Server error')
+        }
+    }
+)
 
 module.exports = router
